Adds section and links about testing facility
Testing analyzers is an important part of the analyzer's development.
Best practices will hopefully evolve and improve over time.
This also means turning the best practices into testing frameworks.

Therefore, a separate section for testing definitely makes sense.

Microsoft.CodeAnalysis.Testing is used in e.g. .NET Analyzers and
the Code Cracker. It's part of the Roslyn SDK and can be seen as a
standard for testing analyzers. RoslynTestKit is rather new but it
provides a lightweight alternative worth exploring further. E.g. its
[| |] syntax.

It is interesting to see what else exists in the wild.
E.g the awesome CSharpGuidelinesAnalyzer uses a small proprietary
framework: https://github.com/bkoelman/RoslynTestFramework

I hope to see this field evolving in the future.